refactor fixtures into conftests.py
